backend/requirements.txt
requests==2.26.0
Django==3.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.2.1
djoser==2.2.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
psycopg2-binary==2.9.3
PyYAML==6.0
python-dotenv==1.0.0
gunicorn==20.1.0
Pillow==9.0.0